requirements.txt
django
django-allauth

python-dateutil
pytz

google-api-python-client
google-auth-httplib2
google-auth-oauthlib
